Convert script to ts and add comment
